GFETCH-119 Refactors OrderModel
