+++ app/src/main/java/co/gobd/gofetch/service/order/SupportedOrderCallback.java
@@ -3,7 +3,7 @@ package co.gobd.gofetch.service.order;
 import java.util.List;
 
 import co.gobd.gofetch.network.ConnectionCallback;
-import co.gobd.gofetch.model.SupportedOrder;
+import co.gobd.gofetch.model.order.SupportedOrder;
 
 /**
  * Created by tonmoy on 19-Apr-16.
